+++ noa-change-detection-mapping/CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.3.0] - 2025-10-21
+### Added
+- Zarr generation for assets and upload to s3 (#16)
+### Changed
+- Assets are now COGs not GTiff
+
 ## [0.2.12] - 2025-06-20
 ### Fixed
 - Another Path bug. Trying to solve kafka bugs (#172)